End of first lesson

+++ FirstScript.sql
@@ -1,8 +1,19 @@
-DROP TABLE FIRMY
-DROP TABLE OSOBY
-DROP TABLE MIASTA
-DROP TABLE WOJ
 
+IF OBJECT_ID(N'ETATY') IS NOT NULL
+	DROP TABLE ETATY
+GO
+IF OBJECT_ID(N'OSOBY') IS NOT NULL
+	DROP TABLE OSOBY
+GO
+IF OBJECT_ID(N'FIRMY') IS NOT NULL
+	DROP TABLE FIRMY
+GO
+IF OBJECT_ID(N'MIASTA') IS NOT NULL
+	DROP TABLE MIASTA
+GO
+IF OBJECT_ID(N'WOJ') IS NOT NULL
+	DROP TABLE WOJ
+GO
 
 
 CREATE TABLE dbo.WOJ 
@@ -10,12 +21,16 @@ CREATE TABLE dbo.WOJ
 , nazwa nvarchar(50) NOT NULL 
 )
 
+GO
+
 CREATE TABLE dbo.MIASTA
 (	id_miasta	int				not null	IDENTITY CONSTRAINT PK_MIASTA PRIMARY KEY
 ,	nazwa		nvarchar(50)	not null	
 ,	kod_woj		nchar(4)		not null	CONSTRAINT FK_MIASTA_WOJ 
 											FOREIGN KEY REFERENCES WOJ(kod_woj)
 )
+
+GO
 
 CREATE TABLE dbo.OSOBY
 (
@@ -28,7 +43,7 @@ CREATE TABLE dbo.OSOBY
 
 GO
 
-CREATE TABLE FIRMY
+CREATE TABLE dbo.FIRMY
 (
 	nazwa_skr		nchar(4)		not null	CONSTRAINT PK_FIRMY PRIMARY KEY
 ,	id_miasta		int				not null	CONSTRAINT FK_FIRMY_MIASTA
@@ -38,7 +53,18 @@ CREATE TABLE FIRMY
 ,	ulica			nvarchar(50)	not null	
 )
 
+go
 
+CREATE TABLE dbo.ETATY
+(
+	id_osoby	int				not null	CONSTRAINT FK_ETATY_OSOBY	FOREIGN KEY REFERENCES OSOBY(id_osoby)
+,	id_firmy	nchar(4)		not null	CONSTRAINT FK_FIRMY			FOREIGN KEY REFERENCES FIRMY(nazwa_skr)
+,	stanowisko	nvarchar(50)	not null	
+,	pensja		MONEY			not null
+,	od			DATETIME		not null
+,	do			DATETIME		not null
+,	id_etatu	int				not null	IDENTITY CONSTRAINT PK_ETATY	PRIMARY KEY
+)
 
 GO
 
@@ -64,6 +90,7 @@ INSERT INTO MIASTA (nazwa, kod_woj)	VALUES (N'Krakow' , N'malo')
 INSERT INTO MIASTA (nazwa, kod_woj)	VALUES (N'Wieliczka' , N'malo')
 
 
+
 INSERT INTO OSOBY (id_miasta, imie, nazwisko)	VALUES (1 , N'Jan', N'Kowalski')
 INSERT INTO OSOBY (id_miasta, imie, nazwisko)	VALUES (4 , N'Adam', N'Nowak')
 INSERT INTO OSOBY (id_miasta, imie, nazwisko)	VALUES (7 , N'Ferdynand', N'Kiepski')
@@ -81,16 +108,23 @@ INSERT INTO OSOBY (id_miasta, imie, nazwisko)	VALUES (2, N'Anna', N'Kafelek')
 
 
 INSERT INTO FIRMY(nazwa_skr ,id_miasta, nazwa, kod_pocztowy, ulica)	VALUES (N'goog', 1, N'Google', N'00001', N'Mazowiecka')
-INSERT INTO FIRMY(nazwa_skr ,id_miasta, nazwa, kod_pocztowy, ulica)	VALUES (N'sams', 1, N'Samsung', N'05800', N'Kierowska')
-INSERT INTO FIRMY(nazwa_skr ,id_miasta, nazwa, kod_pocztowy, ulica)	VALUES (N'appl', 1, N'Apple', N'45573', N'Trebicka')
-INSERT INTO FIRMY(nazwa_skr ,id_miasta, nazwa, kod_pocztowy, ulica)	VALUES (N'goog', 1, N'Google', N'00001', N'Mazowiecka')
-INSERT INTO FIRMY(nazwa_skr ,id_miasta, nazwa, kod_pocztowy, ulica)	VALUES (N'goog', 1, N'Google', N'00001', N'Mazowiecka')
-INSERT INTO FIRMY(nazwa_skr ,id_miasta, nazwa, kod_pocztowy, ulica)	VALUES (N'goog', 1, N'Google', N'00001', N'Mazowiecka')
-INSERT INTO FIRMY(nazwa_skr ,id_miasta, nazwa, kod_pocztowy, ulica)	VALUES (N'goog', 1, N'Google', N'00001', N'Mazowiecka')
-INSERT INTO FIRMY(nazwa_skr ,id_miasta, nazwa, kod_pocztowy, ulica)	VALUES (N'goog', 1, N'Google', N'00001', N'Mazowiecka')
+INSERT INTO FIRMY(nazwa_skr ,id_miasta, nazwa, kod_pocztowy, ulica)	VALUES (N'sams', 2, N'Samsung', N'05800', N'Kierowska')
+INSERT INTO FIRMY(nazwa_skr ,id_miasta, nazwa, kod_pocztowy, ulica)	VALUES (N'oran', 6, N'Orange', N'58160', N'Milosza')
+INSERT INTO FIRMY(nazwa_skr ,id_miasta, nazwa, kod_pocztowy, ulica)	VALUES (N'xiai', 8, N'Xiaiomi', N'62800', N'Tarasowska')
+INSERT INTO FIRMY(nazwa_skr ,id_miasta, nazwa, kod_pocztowy, ulica)	VALUES (N'appl', 3, N'Apple', N'45573', N'Trebicka')
+INSERT INTO FIRMY(nazwa_skr ,id_miasta, nazwa, kod_pocztowy, ulica)	VALUES (N'pkob', 5, N'PKO Bank Polski', N'05822', N'Moniuszki')
+INSERT INTO FIRMY(nazwa_skr ,id_miasta, nazwa, kod_pocztowy, ulica)	VALUES (N'huaw', 4, N'Huawei', N'60001', N'Batorego')
+INSERT INTO FIRMY(nazwa_skr ,id_miasta, nazwa, kod_pocztowy, ulica)	VALUES (N'tmob', 7, N'Tmobile', N'59200', N'Czeslawska')
+
+
+
+INSERT INTO ETATY(id_osoby ,id_firmy, stanowisko, pensja, od, do)	VALUES ('goog', 1, N'Google', N'00001', N'Mazowiecka')
+
 
 
 
 SELECT * FROM WOJ
 SELECT * FROM MIASTA
 SELECT * FROM OSOBY
+SELECT * FROM FIRMY
+SELECT * FROM ETATY
